# DogAdoption/cmd/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "An API with centralized dog adoption information.",
        "title": "DogAdoptionApp REST HATEOAS API",
        "contact": {
            "name": "Gustav Näslund, gn222gq",
            "email": "gn222gq@student.lnu.se"
        },
        "version": "1.0"
    },
    "host": "https://cscloud7-113.lnu.se/dogadoption",
    "basePath": "/api/v1",
    "paths": {
...
        "/dogshelters": {
            "get": {
                "description": "Retrieves a list of dog shelters based on provided query parameters like location and capacity.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dogshelters"
                ],
                "summary": "Get dog shelters",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by name",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by country",
                        "name": "country",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by city",
                        "name": "city",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success, returns a list of dog shelters",
                        "schema": {
                            "$ref": "#/definitions/dogshelterdto.DogSheltersAndPaginationLinksDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request if the query parameters are invalid",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error if an error occurs while processing the request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Adds a new dog shelter to the system with the provided shelter data in JSON format.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dogshelters"
                ],
                "summary": "Add a new dog shelter",
                "parameters": [
                    {
                        "description": "Dog Shelter Data",
                        "name": "shelter",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dogshelterdto.NewDogShelterDTO"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Success, returns the newly created dog shelter information",
                        "schema": {
                            "$ref": "#/definitions/dogshelterdto.DogShelterDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request, if the JSON body cannot be parsed, mandatory fields are missing, or the dog shelter data is incomplete",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized, if the user does not have permission to add a dog shelter",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error, if an error occurs while processing the request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/dogshelters/{id}": {
            "get": {
                "description": "Retrieves detailed information about a specific dog shelter identified by its unique ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dogshelters"
                ],
                "summary": "Get a dog shelter by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Shelter ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success, returns detailed information about the dog shelter",
                        "schema": {
                            "$ref": "#/definitions/dogshelterdto.DogShelterDTO"
                        }
                    },
                    "404": {
                        "description": "Not Found, if no dog shelter matches the provided ID",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error, if an error occurs while processing the request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Updates the information for an existing dog shelter specified by its ID with the provided shelter data in JSON format.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dogshelters"
                ],
                "summary": "Update a dog shelter",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Shelter ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Dog Shelter Update Data",
                        "name": "shelter",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dogshelterdto.UpdateDogShelterDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success, returns the updated dog shelter information",
                        "schema": {
                            "$ref": "#/definitions/dogshelterdto.DogShelterDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request, if the JSON body cannot be parsed, mandatory fields are missing, or the dog shelter data is incomplete",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized, if the user does not have permission to update the dog shelter",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error, if an error occurs while processing the request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Deletes a dog shelter specified by its ID if the requester has the necessary permissions.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dogshelters"
                ],
                "summary": "Delete a dog shelter",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Shelter ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Dog shelter deleted successfully"
                    },
                    "400": {
                        "description": "Bad Request if the request was malformed",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized if the user does not have permission to delete the dog shelter",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found if the dog shelter with the specified ID does not exist",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error for any server errors",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dogshelterdto.DogShelterDTO": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "city": {
                    "type": "string"
                },
                "country": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "links": {
                    "$ref": "#/definitions/dogshelterdto.DogShelterDtoLinks"
                },
                "name": {
                    "type": "string"
                },
                "website": {
                    "type": "string"
                }
            }
        },
        "dogshelterdto.DogShelterDtoLinks": {
            "type": "object",
            "properties": {
                "dogs_link": {
                    "type": "string"
                },
                "self_link": {
                    "type": "string"
                }
            }
        },
        "dogshelterdto.DogSheltersAndPaginationLinksDTO": {
            "type": "object",
            "properties": {
                "dog_shelter_data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dogshelterdto.DogShelterDTO"
                    }
                },
                "pagination_links": {
                    "$ref": "#/definitions/dto.PaginationLinksDTO"
                }
            }
        },
        "dogshelterdto.NewDogShelterDTO": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "city": {
                    "type": "string"
                },
                "country": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                },
                "website": {
                    "type": "string"
                }
            }
        },
        "dogshelterdto.UpdateDogShelterDTO": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "city": {
                    "type": "string"
                },
                "country": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "website": {
                    "type": "string"
                }
            }
        },
...
        "dto.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "dto.PaginationLinksDTO": {
            "type": "object",
            "properties": {
                "first": {
                    "type": "string"
                },
                "last": {
                    "type": "string"
                },
                "next": {
                    "type": "string"
                },
                "previous": {
                    "type": "string"
                },
                "self": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}